+++ vite.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  alias: {
+    // does not work. #later
+    //'/@components': path.resolve(__dirname,'.')    // aim: '@' or '/@/' to point to source root
+
+    '/@/': '/'
+  }
+}